Initial work for the multiplatform cache
* Added `apollo-cache-interceptor` as a separate artifact
* The cache is cache_first only at the moment and provided by
  `ApolloCacheInterceptor`
* Added CacheExecutionContext and `Response.cacheContext` to access
  `fromCache`
* Added `apollo-integration-kotlin`, like `apollo-integration` but
  multiplatform. See CacheInterceptorTest for a test
* Added `apollo-testing-support` to group useful classes for testing.
  Ultimately, this could be exported to end users for mock/testing
purposes

Next steps:
* Implement different fetchers
* Make the cache work with the streaming parser

make compile
