+++ apollo-cache-interceptor/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=apollo-cache-interceptor
+POM_NAME=Apollo Cache Interceptor
+POM_DESCRIPTION=Apollo Cache Interceptor
