+++ apollo-integration-kotlin/src/commonTest/kotlin/com/apollographql/apollo/integration/Utils.kt
@@ -1,0 +1,3 @@
+package com.apollographql.apollo.integration
+
+expect fun fixtureResponse(name: String): String
