--- apollo-integration/src/main/AndroidManifest.xml
@@ -1,7 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.ben_schwab.topshelf.apollointegration">
-
-  <application android:allowBackup="false" android:label="@string/app_name" android:supportsRtl="true">
-
-  </application>
-
-</manifest>
