+++ apollo-runtime-kotlin/src/commonMain/kotlin/com/apollographql/apollo/internal/RealApolloCall.kt
@@ -38,7 +38,9 @@ class RealApolloCall<D : Operation.Data> constructor(
     }.flatMapLatest { interceptorChain ->
       interceptorChain.proceed(request)
     }.map { apolloResponse ->
-      apolloResponse.response
+      apolloResponse.response.copy(
+          executionContext = apolloResponse.executionContext
+      )
     }
   }
 }
