--- apollo-runtime-kotlin/src/commonTest/kotlin/com/apollographql/apollo/TestUtils.kt
@@ -1,7 +1,0 @@
-package com.apollographql.apollo
-
-import kotlinx.coroutines.CoroutineScope
-import kotlin.coroutines.CoroutineContext
-import kotlin.coroutines.EmptyCoroutineContext
-
-expect fun <T> runBlocking(context: CoroutineContext = EmptyCoroutineContext, block: suspend CoroutineScope.() -> T): T
