--- apollo-runtime-kotlin/src/commonTest/kotlin/com/apollographql/apollo/mock/TestLoggerExecutor.kt
@@ -1,29 +1,0 @@
-package com.apollographql.apollo.mock
-
-import com.apollographql.apollo.api.ApolloExperimental
-import com.apollographql.apollo.api.Operation
-import com.apollographql.apollo.interceptor.ApolloInterceptorChain
-import com.apollographql.apollo.interceptor.ApolloRequest
-import com.apollographql.apollo.interceptor.ApolloRequestInterceptor
-import com.apollographql.apollo.interceptor.ApolloResponse
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.flow.onStart
-
-@ApolloExperimental
-@ExperimentalCoroutinesApi
-object TestLoggerExecutor : ApolloRequestInterceptor {
-
-  override fun <D : Operation.Data> intercept(request: ApolloRequest<D>, chain: ApolloInterceptorChain): Flow<ApolloResponse<D>> {
-    println("Preparing `${request.operation.name().name()}` GraphQL operation for execution ...")
-    return chain.proceed(request)
-        .onStart { println("Started `${request.operation.name().name()}` GraphQL operation execution ...") }
-        .onEach { response -> println("Finished `${request.operation.name().name()}` GraphQL operation execution, response: $response") }
-        .catch { e ->
-          println("Failed `${request.operation.name().name()}` GraphQL operation execution due to error: $e")
-          throw e
-        }
-  }
-}
