--- apollo-runtime-kotlin/src/commonTest/kotlin/com/apollographql/apollo/mock/TestTokenProvider.kt
@@ -1,16 +1,0 @@
-package com.apollographql.apollo.mock
-
-import com.apollographql.apollo.interceptor.TokenProvider
-
-class TestTokenProvider(currentAccessToken: String,
-                        val newAccessToken: String) : TokenProvider {
-  var accessToken = currentAccessToken
-  override suspend fun currentToken(): String {
-    return accessToken
-  }
-
-  override suspend fun refreshToken(previousToken: String): String {
-    accessToken = newAccessToken
-    return accessToken
-  }
-}
