--- apollo-runtime-kotlin/src/iosSimTest/kotlin/com/apollographql/apollo/IOSTestUtils.kt
@@ -1,8 +1,0 @@
-package com.apollographql.apollo
-
-import kotlinx.coroutines.CoroutineScope
-import kotlin.coroutines.CoroutineContext
-
-actual fun <T> runBlocking(context: CoroutineContext, block: suspend CoroutineScope.() -> T): T {
-  return kotlinx.coroutines.runBlocking(context, block)
-}
