+++ apollo-runtime-kotlin/src/iosSimTest/kotlin/com/apollographql/apollo/network/http/ApolloHttpNetworkTransportTest.kt
@@ -6,13 +6,13 @@ import com.apollographql.apollo.api.ApolloExperimental
 import com.apollographql.apollo.api.ExecutionContext
 import com.apollographql.apollo.api.ScalarTypeAdapters
 import com.apollographql.apollo.interceptor.ApolloRequest
-import com.apollographql.apollo.mock.MockQuery
+import com.apollographql.apollo.testing.MockQuery
 import com.apollographql.apollo.network.HttpExecutionContext
 import com.apollographql.apollo.network.HttpMethod
 import com.apollographql.apollo.network.mock.MockHttpResponse
 import com.apollographql.apollo.network.mock.MockSessionDataTask
 import com.apollographql.apollo.network.toNSData
-import com.apollographql.apollo.runBlocking
+import com.apollographql.apollo.testing.runBlocking
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.single
 import okio.ByteString.Companion.encodeUtf8
