--- apollo-runtime-kotlin/src/jvmTest/kotlin/com/apollographql/apollo/JvmTestUtils.kt
@@ -1,8 +1,0 @@
-package com.apollographql.apollo
-
-import kotlinx.coroutines.CoroutineScope
-import kotlin.coroutines.CoroutineContext
-
-actual fun <T> runBlocking(context: CoroutineContext, block: suspend CoroutineScope.() -> T): T {
-  return kotlinx.coroutines.runBlocking(context, block)
-}
