+++ apollo-testing-support/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=apollo-testing-support
+POM_NAME=Apollo testing support
+POM_DESCRIPTION=Apollo testing support
