+++ apollo-testing-support/src/commonMain/kotlin/com/apollographql/apollo/testing/TestTokenProvider.kt
@@ -1,0 +1,16 @@
+package com.apollographql.apollo.testing
+
+import com.apollographql.apollo.interceptor.TokenProvider
+
+class TestTokenProvider(currentAccessToken: String,
+                        val newAccessToken: String) : TokenProvider {
+  var accessToken = currentAccessToken
+  override suspend fun currentToken(): String {
+    return accessToken
+  }
+
+  override suspend fun refreshToken(previousToken: String): String {
+    accessToken = newAccessToken
+    return accessToken
+  }
+}
