+++ apollo-testing-support/src/commonMain/kotlin/com/apollographql/apollo/testing/TestUtils.kt
@@ -1,0 +1,7 @@
+package com.apollographql.apollo.testing
+
+import kotlinx.coroutines.CoroutineScope
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
+
+expect fun <T> runBlocking(context: CoroutineContext = EmptyCoroutineContext, block: suspend CoroutineScope.() -> T): T
