+++ apollo-testing-support/src/iosMain/kotlin/com/apollographql/apollo/testing/IOSTestUtils.kt
@@ -1,0 +1,8 @@
+package com.apollographql.apollo.testing
+
+import kotlinx.coroutines.CoroutineScope
+import kotlin.coroutines.CoroutineContext
+
+actual fun <T> runBlocking(context: CoroutineContext, block: suspend CoroutineScope.() -> T): T {
+  return kotlinx.coroutines.runBlocking(context, block)
+}
