+++ composite/build.gradle.kts
@@ -10,6 +10,7 @@ buildscript {
   }
 
   dependencies {
+    classpath(groovy.util.Eval.x(project, "x.dep.oneEightPlugin"))
     classpath(groovy.util.Eval.x(project, "x.dep.android.plugin"))
     classpath(groovy.util.Eval.x(project, "x.dep.kotlin.plugin"))
     classpath("com.apollographql.apollo:apollo-gradle-plugin")
