+++ gradle/dependencies.gradle
@@ -19,6 +19,7 @@ def versions = [
     okHttp4               : '4.9.0',
     okHttp                : '3.12.11', // Keep this to keep supporting older Android devices
     okio                  : '2.9.0',
+    okioSnapshot          : '2.10.0-SNAPSHOT',
     rxjava                : '2.2.20',
     rxjava3               : '3.0.7',
     rxandroid             : '2.0.1',
@@ -51,6 +52,7 @@ ext.dep = [
     cache                 : "com.nytimes.android:cache:$versions.cache",
     gradleJapiCmpPlugin   : "me.champeau.gradle:japicmp-gradle-plugin:0.2.8",
     gradleMetalavaPlugin  : "me.tylerbwong.gradle:metalava-gradle:0.1.5",
+    oneEightPlugin        : "gradle.plugin.net.mbonnin:one-eight:0.1",
     vanniktechPlugin      : "com.vanniktech:gradle-maven-publish-plugin:0.11.1",
     gradlePublishPlugin   : "com.gradle.publish:plugin-publish-plugin:0.12.0",
     guavaJre              : "com.google.guava:guava:$versions.guava",
@@ -77,6 +79,7 @@ ext.dep = [
         logging4      : "com.squareup.okhttp3:logging-interceptor:$versions.okHttp4",
     ],
     okio                  : "com.squareup.okio:okio-multiplatform:$versions.okio",
+    okioFiles          : "com.squareup.okio:okio-files:$versions.okioSnapshot",
     poet                  : [
         java  : "com.squareup:javapoet:$versions.javaPoet",
         kotlin: "com.squareup:kotlinpoet:$versions.kotlinPoet",
